+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Removes support for iOS 12
+- Removes CryptoSwift dependency
+
 ## [0.4.0] - 2021-03-16
 ### Changed
 - Now distributing XCFramework instead of fat one